[BUGFIX] Fix wrong be module identifier
In ext_tables was "cutemailing" the identifier and in
`Backend/Modules.php` was it "newsletter"
In be_groups module can't be allowed for editors. It was not visible in
the list. After change it to "cutemailing" also in `Backend/Modules.php`
everything was fine.

I'm not sure if that is a breaking change. The identifier is just needed
by TYPO3 to find the right module and for allow in access rights
for editors. But with that wrong identifier no newsletter module can't
be allowed. Therefore it should be no problem and no breaking change.
But mabe I'm wrong. Not sure. Has to be checked by another dev before
release the new version.

Related: #23404